+++ Core/Commands/JumpNotZeroCommand.cs
@@ -26,8 +26,8 @@ namespace ProcessorEmulator.Commands
         {
             if (processor.ZeroFlag)
                 return;
-            uint value = processor.GetOperandValue(AddressingTwo, OperandTwo);
-            processor.commandPointer = value;
+            int value = processor.GetOperandValue(AddressingTwo, OperandTwo);
+            processor.commandPointer = (uint)value;
         }
     }
 }
